app/python/requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
